// frontend/src/types/fileResponseStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type FileResponseStatus = typeof FileResponseStatus[keyof typeof FileResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FileResponseStatus = {
  PENDING: 'PENDING',
  ACTIVE: 'ACTIVE',
  REJECTED: 'REJECTED',
} as const;